SCORM API
Added Scorm functionality through the Scorm class. Each time a bone is updated, the status and score are set and submitted to the LMS.
Each bone is 3 points, a mistake is -1 point. Maximum points is 100%. 50%+ is a passing grade.
Issues: lagspike every time you place a bone. Time is not being set to the LMS. Only the final score (ant not the result for each bone) is being recorded.