Updated scanner
Changed it so that each bone changes colour individually when being scanned, so it's easier to tell which one it is that you're scanning.
Unfortunately, this requires each bone to have a separate unique material, so the game size balloons out by about 50%. I could try reducing it, but removing the scanner would remove a lot of issues altogether.